feat: added all the types into the specific type file
