+++ src/types/Attestation.ts
@@ -5,6 +5,7 @@
 import ICType from './CType'
 import IPublicIdentity from './PublicIdentity'
 import { IDelegationBaseNode } from './Delegation'
+import Attestation from '../attestation/Attestation'
 
 export default interface IAttestation {
   claimHash: string
@@ -13,3 +14,11 @@ export default interface IAttestation {
   delegationId: IDelegationBaseNode['id'] | null
   revoked: boolean
 }
+
+export type CompressedAttestation = [
+  Attestation['claimHash'],
+  Attestation['cTypeHash'],
+  Attestation['owner'],
+  Attestation['revoked'],
+  Attestation['delegationId']
+]
