+++ src/types/AttestedClaim.ts
@@ -2,10 +2,18 @@
  * @packageDocumentation
  * @module IAttestedClaim
  */
-import IRequestForAttestation from './RequestForAttestation'
-import IAttestation from './Attestation'
+
+import IRequestForAttestation, {
+  CompressedRequestForAttestation,
+} from './RequestForAttestation'
+import IAttestation, { CompressedAttestation } from './Attestation'
 
 export default interface IAttestedClaim {
   attestation: IAttestation
   request: IRequestForAttestation
 }
+
+export type CompressedAttestedClaim = [
+  CompressedRequestForAttestation,
+  CompressedAttestation
+]
