+++ src/types/Claim.ts
@@ -10,3 +10,11 @@ export default interface IClaim {
   cTypeHash: ICType['hash']
   owner: IPublicIdentity['address']
 }
+
+export type CompressedClaimContents = object
+
+export type CompressedClaim = [
+  CompressedClaimContents,
+  IClaim['cTypeHash'],
+  IClaim['owner']
+]
